requirements.txt
# DataBot - Production Dependencies
# Install with: pip install -r requirements.txt

# Core Framework
streamlit==1.31.0

# LLM & AI
langchain-groq==0.1.0
langchain==0.1.0
openai==1.12.0  # Optional: for GPT models
anthropic==0.18.0  # Optional: for Claude

# File Processing
pypdf==4.0.1
python-docx==1.1.0  # Word documents
openpyxl==3.1.2  # Excel files
pillow==10.2.0  # Image processing
pytesseract==0.3.10  # OCR for scanned documents

# Database
psycopg2-binary==2.9.9  # PostgreSQL
sqlalchemy==2.0.25  # ORM (optional)
alembic==1.13.1  # Database migrations

# Security & Authentication
bcrypt==4.1.2  # Password hashing
pyjwt==2.8.0  # JWT tokens
python-dotenv==1.0.1  # Environment variables
cryptography==42.0.2  # Encryption

# API & Web
fastapi==0.109.2  # REST API (optional)
uvicorn==0.27.1  # ASGI server
requests==2.31.0  # HTTP client
httpx==0.26.0  # Async HTTP

# Background Tasks
celery==5.3.6  # Task queue
redis==5.0.1  # Message broker

# Monitoring & Logging
sentry-sdk==1.40.0  # Error tracking
python-json-logger==2.0.7  # Structured logging

# Data Validation
pydantic==2.6.1  # Data validation
email-validator==2.1.0  # Email validation

# Utilities
pandas==2.2.0  # Data manipulation
numpy==1.26.4  # Numerical operations
python-dateutil==2.8.2  # Date parsing

# Email
sendgrid==6.11.0  # Email service
python-multipart==0.0.9  # File uploads

# Cloud Storage (Optional)
boto3==1.34.34  # AWS S3
google-cloud-storage==2.14.0  # Google Cloud Storage

# Payment Processing (Optional)
stripe==8.2.0  # Payment processing

# Development Tools
pytest==8.0.0  # Testing
pytest-cov==4.1.0  # Test coverage
black==24.1.1  # Code formatting
flake8==7.0.0  # Linting
mypy==1.8.0  # Type checking

# Production Server
gunicorn==21.2.0  # WSGI server
